+++ Service/QueryBuilderManipulator.php
@@ -63,8 +63,20 @@ class QueryBuilderManipulator
 		}
 		
 		if ($qb instanceof QueryBuilder) {
-			$qb->andWhere(isset($expr) ? $expr : (new \Doctrine\ORM\Query\Expr())->eq($alias.'.'.$field, $namedParameter));
-			$qb->setParameter($namedParameter, $value);
+			if (!isset($expr)) {
+				$expr = (new \Doctrine\ORM\Query\Expr())->eq($alias.'.'.$field, $namedParameter);
+				$qb->setParameter($namedParameter, $exprValue);
+			} else {
+				if (is_array($value)) {
+					foreach ($value as $key => $val) {
+						$qb->setParameter($key, $val);
+					}
+				} else {
+					$qb->setParameter($namedParameter, $value);
+				}
+			}
+			
+			$qb->andWhere($expr);
 		} else {
 			$qb->addAnd(isset($expr) ? $expr : (new \Doctrine\MongoDB\Query\Expr())->field($field)->equals($value));
 		}